Update login method and add new api to find host and user by account_id

+++ exe-201-beestay-be/exe-201-beestay-be/src/main/java/com/exe201/project/exe_201_beestay_be/dto/responses/StayCationDetailResponse.java
@@ -13,7 +13,7 @@ public class StayCationDetailResponse {
     private LocationResponse location;
     private double distanceToCenter;
 
-    private double averageRating;
+    private float averageRating;
     private long reviewCount;
     private List<ReviewResponse> reviews;
